Add support for material tags in config
This removes a number of individually defined materials in favour of
block tags, which are dynamically updated by Mojang (such as when new
tree variants are added).

This cuts down on unnecessary manual config updates for game updates.

For now, flowers are not included as they can cause duplicate drops,
and restoring double-high plants results in loss of block state (eg top).